* massive restructuring for activity module
