+++ public/modules/autoload.php
@@ -1,23 +1,25 @@
 <?php 
-require_once('login/Login.php'); // (*) requires db
-require_once('profile/Profile.php'); // (*) requires db
-require_once('login/Register.php'); // (*) requires db
-require_once('settings/Settings.php'); // (*)
+require_once('login/Login.php'); //
+require_once('profile/Profile.php'); //
+require_once('login/Register.php'); //
+require_once('settings/Settings.php'); //
 /*
- * Site specific Modules.
+ * All these require one or all of the modules listed above :
  */
-require_once('identity/Identity.php'); // (*)
-require_once('activity/Activity.php'); // (*)
-require_once('activity/Comment.php'); // (*)
-require_once('activity/Activity_Shout.php'); // (*)
-require_once('activity/Activity_Event.php'); // (*)
-require_once('activity/Activity_Event_Widgets.php'); // (*)
-require_once('activity/Activity_Poll.php'); // (*)
-require_once('activity/Activity_ActionMsg.php'); // (*)
-require_once('pages/Pages.php'); // (*)
-require_once('gw2api/gw2api.php'); // (*)
-require_once('gw2api/gw2api_Widgets.php'); // (*)
+require_once('identity/Identity.php'); //
+require_once('activity/Activity.php'); //
+require_once('activity/Comment.php'); //
+require_once('activity/Activity_Shout.php'); //
+require_once('activity/event/Activity_Event.php'); //
+require_once('activity/Event/Activity_Event_Roles.php'); //
+require_once('activity/event/Activity_Event_Signups.php'); //
+require_once('activity/event/Activity_Event_Widgets.php'); //
+require_once('activity/Activity_Poll.php'); //
+require_once('activity/Activity_ActionMsg.php'); //
+require_once('pages/Pages.php'); //
+require_once('gw2api/gw2api.php'); //
+require_once('gw2api/gw2api_Widgets.php'); //
 /*
  * for development purposes
  */
-require_once('dev/Dev.php'); // (*)
+require_once('dev/Dev.php'); //
