+++ public/themes/boilerplate/views/activity/event/activity_event_signup_button.php
@@ -1,3 +1,4 @@
 <form method="post" action="{##signup##}">
     <input type="submit" class="button" name="activity[signup]" value="{##signup_text##}" />
+    <input type="hidden" name="target_url" value="{##target_url##}" />
 </form>
